requirements.txt
click
Django
django-tinymce
funcsigs
gunicorn==19.4.0
Jinja2
MarkupSafe
mock
mysql
pbr
Pillow
pytz
six
virtualenv
Werkzeug
whitenoise
dj-database-url==0.2.1
django-storages
dropbox